+++ main/render.h
@@ -93,10 +93,10 @@ void render_start_frame(void);
 
 // Given a list of point numbers, rotate any that haven't been rotated
 // this frame
-g3s_codes rotate_list(int nv, short *pointnumlist);
+g3s_codes rotate_list(int nv, int *pointnumlist);
 
 // Given a list of point numbers, project any that haven't been projected
-void project_list(int nv, short *pointnumlist);
+void project_list(int nv, int *pointnumlist);
 
 extern void render_mine(int start_seg_num, fix eye_offset, int window_num);
